requirements.txt
dnspython
marshmallow
marshmallow_dataclass
python-dateutil
cement==3.0.4
jinja2
pyyaml
colorlog
semver
jwt
iso8601
tabulate
pyOpenSSL
requests
